// server/@types/make-recall-decision-api/models/ManagerRecallDecisionTypeSelectedValue.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type ManagerRecallDecisionTypeSelectedValue = {
    value?: ManagerRecallDecisionTypeSelectedValue.value;
    details?: string;
};

export namespace ManagerRecallDecisionTypeSelectedValue {

    export enum value {
        NO_RECALL = 'NO_RECALL',
        RECALL = 'RECALL',
    }


}